move test to model folder
