+++ spec/model/repeat-builder.spec.ts
@@ -14,31 +14,33 @@
  * limitations under the License.
  *
  */
+import {RepeatBuilder} from '../../src/model/repeat.builder';
 
-import {ActionDataFilterBuilder} from '../src/model/action-data-filter.builder';
-
-describe("ActionDataFilter", () => {
+describe('repeat builder', () => {
 	
-	it("should generate an empty object", () => {
-		expect(new ActionDataFilterBuilder().build()).toEqual({});
-	});
-	
-	it("should generate a populated object", () => {
-		expect(new ActionDataFilterBuilder()
-			.withFromStateData("fromState")
-			.withToStateData("toState")
-			.withResults("result")
-			.build()).toEqual(
-			{
-				fromStateData: "fromState",
-				toStateData: "toState",
-				results: "result",
-			},
-		);
+	it("Should create empty object", () => {
 		
+		expect(new RepeatBuilder().build()).toEqual({});
 		
 	});
 	
+	it("should create an object with all possible fields", () => {
+		
+		expect(new RepeatBuilder()
+			.withExpression("anyExpression")
+			.withCheckBefore(true)
+			.withMax(3)
+			.withContinueOnError(true)
+			.withStopOnEvents(["onStopEvent"])
+			.build())
+			.toEqual({
+				expression: "anyExpression",
+				checkBefore: true,
+				max: 3,
+				continueOnError: true,
+				stopOnEvents: ["onStopEvent"],
+			});
+		
+	});
 	
 });
-
